requirement.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
discord.py==1.6.0
idna==3.1
meteo-downloader==1.0
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.16.0
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
